refactor(modules): move internal modules
