--- internal/api/cronjob/status_handler.go
@@ -1,24 +1,0 @@
-package cronjob
-
-import (
-	"github.com/darchlabs/synchronizer-v2/internal/api"
-	"github.com/gofiber/fiber/v2"
-)
-
-func statusCronjobHandler(ctx Context) func(c *fiber.Ctx) error {
-	return func(c *fiber.Ctx) error {
-		// get cronjob status
-		status := ctx.Cronjob.GetStatus()
-
-		// prepare response
-		response := struct{
-			status string
-		}{
-			status: status,
-		}
-
-		return c.Status(fiber.StatusOK).JSON(api.Response{
-			Data: response,
-		})
-	}
-}
